+++ blender_batch_exr/__init__.py
@@ -1,2 +1,2 @@
 """Standalone Blender multilayer EXR to Photoshop converter."""
-__version__ = "0.3.0"
+__version__ = "0.4.0"